Refactor text mining to new expandable sections
